backend/requirements.txt
supabase
python-dotenv
fastapi[standard]
aiohttp
pandas
requests